requirements.txt
groq
pydantic
python-dotenv
requests